+++ ruby/lib/natsuzora.rb
@@ -6,19 +6,21 @@ require_relative 'natsuzora/token'
 require_relative 'natsuzora/validator'
 require_relative 'natsuzora/html_escape'
 require_relative 'natsuzora/value'
+require_relative 'natsuzora/data_normalizable'
 require_relative 'natsuzora/ast'
 require_relative 'natsuzora/lexer'
 require_relative 'natsuzora/parser'
 require_relative 'natsuzora/context'
 require_relative 'natsuzora/template_loader'
 require_relative 'natsuzora/renderer'
+require_relative 'natsuzora/payload'
 require_relative 'natsuzora/template'
 require_relative 'natsuzora/contract'
 
 module Natsuzora
   class << self
     def render(source, data, include_root: nil)
-      Template.new(source, include_root: include_root).render(data)
+      Template.new(source, include_root: include_root).render(Payload.new(data))
     end
 
     def parse(source, include_root: nil)